add rowunitprice update
