+++ packages/kokoas-client/src/pages/projEstimateV2/hooks/useSmartHandlers.tsx
@@ -10,9 +10,9 @@ export type UseSmartHandlers =  ReturnType<typeof useSmartHandlers>;
 export const useSmartHandlers = () => {
   const { setValue, getValues } = useFormContext<TypeOfForm>();
 
-  /** 原価の変更 */
+  /***************
+   * 原価の変更 */
   const handleChangeCostPrice = useCallback((rowIdx: number ) => {
-
     const profitRate = getValues(getItemsFieldName<'items.0.materialProfRate'>(rowIdx, 'materialProfRate')) / 100;
     const taxRate = getValues('taxRate') / 100;
     const isTaxable = getValues(getItemsFieldName<'items.0.taxable'>(rowIdx, 'taxable'));
@@ -37,10 +37,12 @@ export const useSmartHandlers = () => {
 
   }, [getValues, setValue]);
 
-  /** 数量の変更 */
+  /***************
+   * 数量の変更 */
   const handleChangeQuantity = useCallback(handleChangeCostPrice, [handleChangeCostPrice]);
 
-  /** 利益率の変更 */
+  /*****************
+   * 利益率の変更 */
   const handleChangeProfitRate = useCallback((rowIdx: number) => {
     const profitRate = getValues(getItemsFieldName<'items.0.materialProfRate'>(rowIdx, 'materialProfRate')) / 100;
     const taxRate = getValues('taxRate') / 100;
@@ -65,19 +67,14 @@ export const useSmartHandlers = () => {
   }, [getValues, setValue]);
 
 
-  /** 単価の変更 */
+  /****************
+   * 単価の変更 */
   const handleChangeUnitPrice = useCallback((rowIdx: number) => {
-
-
     const unitPrice = getValues(getItemsFieldName<'items.0.unitPrice'>(rowIdx, 'unitPrice'));
-    console.log(unitPrice);
-
-
     const taxRate = getValues('taxRate') / 100;
     const isTaxable = getValues(getItemsFieldName<'items.0.taxable'>(rowIdx, 'taxable'));
     const costPrice = getValues(getItemsFieldName<'items.0.costPrice'>(rowIdx, 'costPrice'));
     const quantity = getValues(getItemsFieldName<'items.0.quantity'>(rowIdx, 'quantity'));
-
     const {
       rowUnitPriceAfterTax,
       profitRate,
@@ -92,6 +89,30 @@ export const useSmartHandlers = () => {
     setValue(getItemsFieldName<'items.0.materialProfRate'>(rowIdx, 'materialProfRate'), roundTo(profitRate * 100, 2));
     setValue(getItemsFieldName<'items.0.rowUnitPriceAfterTax'>(rowIdx, 'rowUnitPriceAfterTax'), rowUnitPriceAfterTax);
 
+  }, [getValues, setValue]);
+
+  /************************
+   * 金額（税込み）の変更 */
+  const handleChangeRowUnitPriceAfterTax = useCallback((rowIdx: number)=>{
+    const rowUnitPriceAfterTax = getValues(getItemsFieldName<'items.0.rowUnitPriceAfterTax'>(rowIdx, 'rowUnitPriceAfterTax'));
+    const taxRate = getValues('taxRate') / 100;
+    const isTaxable = getValues(getItemsFieldName<'items.0.taxable'>(rowIdx, 'taxable'));
+    const costPrice = getValues(getItemsFieldName<'items.0.costPrice'>(rowIdx, 'costPrice'));
+    const quantity = getValues(getItemsFieldName<'items.0.quantity'>(rowIdx, 'quantity'));
+
+    const {
+      unitPrice,
+      profitRate,
+    } = calculateEstimateRow({
+      rowUnitPriceAfterTax,
+      costPrice,
+      quantity,
+      taxRate,
+      isTaxable,
+    });
+
+    setValue(getItemsFieldName<'items.0.materialProfRate'>(rowIdx, 'materialProfRate'), roundTo(profitRate * 100, 2));
+    setValue(getItemsFieldName<'items.0.unitPrice'>(rowIdx, 'unitPrice'), unitPrice);
 
   }, [getValues, setValue]);
 
@@ -100,5 +121,6 @@ export const useSmartHandlers = () => {
     handleChangeQuantity,
     handleChangeProfitRate,
     handleChangeUnitPrice,
+    handleChangeRowUnitPriceAfterTax,
   };
 };
