+++ packages/kokoas-client/src/pages/projEstimateV2/tables/estimate/rowFields/RowUnitPriceAfterTax.tsx
@@ -1,15 +1,16 @@
-import { MoneyField } from 'kokoas-client/src/components/reactHookForm/MoneyField';
+import { OutlinedMoneyInput } from 'kokoas-client/src/components/reactHookForm/OutlinedMoneyInput';
 import { useFormContext, useWatch } from 'react-hook-form';
 import { getItemsFieldName, TypeOfForm } from '../../../form';
+import { UseSmartHandlers } from '../../../hooks/useSmartHandlers';
 
 export const RowUnitPriceAfterTax = ({
   rowIdx,
+  handleChange,
 }: {
-  rowIdx: number
+  rowIdx: number,
+  handleChange: UseSmartHandlers['handleChangeRowUnitPriceAfterTax']
 }) => {
-
-
-  const { control } = useFormContext<TypeOfForm>();
+  const { control, register } = useFormContext<TypeOfForm>();
 
   const [
     costPrice,
@@ -22,17 +23,14 @@ export const RowUnitPriceAfterTax = ({
     control,
   });
 
-
   return (
-    <MoneyField
-      controllerProps={{
-        name: getItemsFieldName(rowIdx, 'rowUnitPriceAfterTax'),
-        control,
-      }}
-      textFieldProps={{
-        size: 'small',
-        disabled: !!envStatus || !(+costPrice),
-      }}
+    <OutlinedMoneyInput {...register(
+      getItemsFieldName(rowIdx, 'rowUnitPriceAfterTax'),
+      {
+        onChange: handleChange,
+        disabled: !!envStatus || !+(costPrice ?? 0),
+      })
+    }
     />
   );
 };
